+++ OrganizationSearch/OrganizationsListFilter/index.js
@@ -1,1 +1,2 @@
 export { default as OrganizationsListFilter } from './OrganizationsListFilter';
+export { TypeFilter } from './TypeFilter';