+++ test/jest/fixtures/organizationTypes.js
@@ -1,0 +1,12 @@
+export const organizationTypes = [
+  {
+    id: 'e7e9af00-c12c-448f-8ad1-d15ff209605a',
+    name: 'Book trade',
+    status: 'Inactive',
+  },
+  {
+    id: 'f04c7277-0019-43cf-84b3-02d894a9d81a',
+    name: 'Auction house',
+    status: 'Active',
+  },
+];